+++ CHANGELOG.md
@@ -6,6 +6,8 @@ TBD - TBD
 	ResultRange doesn't get invalidated upon reconnect. (@Abscissa)
 - **Fixed:** Remove unnecessary precondition that connection must be
 	closed before calling `Connection.connect`. (@Abscissa)
+- **Fixed:** [#121](https://github.com/mysql-d/mysql-native/issues/121):
+	Use `const(char[])` instead `string` for queries. (@Abscissa)
 - **Tests:** Add tests for reconnecting. (@Abscissa)
 - **Internal:** [#144](https://github.com/mysql-d/mysql-native/issues/144):
 	Move all packet handling, etc., into mysql.protocol.comms (@Abscissa)